refactor: old and new setting in side panel
