+++ packages/sn-filter-pane/core/package.json
@@ -1,7 +1,7 @@
 {
   "module": "esm/index.js",
   "peerDependencies": {
-    "@nebula.js/stardust": "^3.1.0-alpha.0",
+    "@nebula.js/stardust": "3.1.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   }
