+++ packages/sn-listbox/package.json
@@ -39,10 +39,10 @@
     "@emotion/react": "11.10.4",
     "@emotion/styled": "11.10.4",
     "@mui/material": "5.10.4",
-    "@nebula.js/cli": "^3.1.0-alpha.2",
-    "@nebula.js/cli-build": "^3.1.0-alpha.2",
-    "@nebula.js/cli-sense": "^3.1.0-alpha.2",
-    "@nebula.js/cli-serve": "^3.1.0-alpha.2",
+    "@nebula.js/cli": "3.1.3",
+    "@nebula.js/cli-build": "3.1.3",
+    "@nebula.js/cli-sense": "3.1.3",
+    "@nebula.js/cli-serve": "3.1.3",
     "@rollup/plugin-typescript": "^8.4.0",
     "@types/enigma.js": "^2.2.2",
     "@types/qlik-engineapi": "^12.67.12",
@@ -67,6 +67,6 @@
     "zustand": "^4.1.1"
   },
   "peerDependencies": {
-    "@nebula.js/stardust": "^3.1.0-alpha.2"
+    "@nebula.js/stardust": "3.1.3"
   }
 }
